+++ HISTORY.rst
@@ -137,3 +137,15 @@ History
 
 * add min confidence when displaying prediction from a segmentation mask model
 * now semantic coco accepts only logits to create annotations
+
+
+0.9.39
+------------
+
+* add new method to remap category idxs
+
+
+0.9.48
+------------
+
+* add new feature to save images and masks to a folder and filter out images and mask with less than k annotations